requirements.txt

python-dotenv
openai
pandas
transformers
torch
tensorflow
tf-keras
tiktoken
sentencepiece
matplotlib
seaborn